refactor(shared): split map.ts into world/ subdirectory (#1562)
Split packages/shared/src/map.ts (3423 lines) into a world/ subdirectory
with 8 focused files, each under 1000 lines:

  world/
    index.ts              # barrel re-export
    map-geometry.ts       # 244 — coordinate helpers + encode/decode (pure)
    pathfinding.ts        # 462 — A*, movement planning, traversability
    fog-of-war.ts         # 266 — visibility + player-view construction
    world-builders.ts     # 933 — constructors, cloners, neutral-AI, hero progression
    action-prediction.ts  # 642 — predictPlayerWorldAction
    tile-mutations.ts     # 811 — validate / precheck / resolve / apply
    battle-outcome.ts     # 249 — applyBattleOutcomeToWorld

Delete the original packages/shared/src/map.ts; shared/index.ts now
re-exports from ./world/index.ts so downstream imports are unaffected
(@veil/shared named exports remain stable).

Extraction strategy:
- Pure function only — no runtime-behavior change; barrel preserves the
  full public API surface.
- Introduced internal-only cross-file imports to keep prior file-local
  helpers encapsulated. Helpers that were previously file-local and are
  now called across layers (clonePosition, findHero, etc.) were exported.
- Broke a potential circular dep between pathfinding and tile-mutations
  by moving clonePosition to map-geometry and inlining a 3-line findHero
  in pathfinding.

Verification:
- npm run typecheck:ci — shared/server/client/cocos all green (modulo the
  2 pre-existing admin-console + auth errors on main, unrelated)
- npm run lint:arch — 0 errors (17 pre-existing circular warnings)
- packages/shared/src/__tests__/map.test.ts — 10/10 pass
- apps/server/test/{fog-of-war-culling,colyseus-large-map-sync,
  ws-action-rate-limit}.test.ts — 4/4 pass

Follow-up: the Cocos project-shared/map.ts mirror is intentionally out of
scope here and will be addressed by #1560 (@veil/shared alias unification).

Closes #1562
